Mostly Finished Rat Rig
Finished setting up the handles, Last parts I need to do with this as far as I am aware is to finish linking and parenting all the controls and skinning everything. Also with this I had to go and remake the tail so it was straight, so this might be an issue if the texture for the rat has already been made and the previous rat was already UV mapped.